delete company

+++ app/Modules/Company/Controllers/CompanyController.php
@@ -118,8 +118,14 @@ class CompanyController extends Controller
     public function destroy($id)
     {
         $company = Company::find($id);
+        if($company)
+        {
         $company->delete();
-        return redirect('/');
+        alert()->success('Succés!', 'La societé a été supprimé avec succés ');
+        return redirect()->back();
+        }
+        alert()->error('Succés!', 'Societé introuvable !');
+        return redirect()->back();
     }
 
 }
